[scanner] 🌱 refactor: split oversized source files (#20967)
Split handlers.compliance.ts (1151 lines) into 4 focused sub-modules,
each under 500 lines, to address the auto-QA large-file finding.

| File | Lines | Responsibility |
|------|-------|----------------|
| handlers.compliance.frameworks.ts | 390 | Frameworks, HIPAA, GxP, BAA, SOD, Change Control, Data Residency |
| handlers.compliance.gov.ts        | 295 | NIST 800-53, DISA STIG, Air-Gap, FedRAMP, Identity & Access |
| handlers.compliance.security.ts   | 315 | SIEM, Incident Response, Threat Intel, Supply Chain (Epic 6) |
| handlers.compliance.erm.ts        | 197 | Enterprise Risk Management (Epic 7) + card templates |
| handlers.compliance.ts (barrel)   |  20 | Re-exports createComplianceHandlers() — public API unchanged |

All consumers continue to import createComplianceHandlers() from the
same path (./handlers.compliance) with no changes required.
